# DeBroglieProject/docs/swagger.yaml
basePath: /api
definitions:
...
  ds.RequestDeBroglieCalculation:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      creator_id:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      moderator_id:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/ds.RequestStatus'
    type: object
  ds.RequestStatus:
    enum:
    - черновик
    - удалён
    - сформирован
    - завершён
    - отклонён
    type: string
    x-enum-varnames:
    - RequestStatusDraft
    - RequestStatusDeleted
    - RequestStatusFormed
    - RequestStatusCompleted
    - RequestStatusRejected
...
  handler.deBroglieCalculationDetail:
    properties:
      de_broglie_length:
        type: number
      id:
        type: integer
      particle_id:
        type: integer
      particle_image:
        type: string
      particle_mass:
        type: number
      particle_name:
        type: string
      speed:
        type: number
    type: object
  handler.deBroglieDraftInfoResponse:
    properties:
      draft_id:
        type: integer
      particles_cnt:
        type: integer
    type: object
  handler.deBroglieRequestDetailResponse:
    properties:
      debrogliecalculations:
        items:
          $ref: '#/definitions/handler.deBroglieCalculationDetail'
        type: array
      completed_at:
        type: string
      created_at:
        type: string
      formed_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  handler.errorResponse:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
...
  handler.successResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для расчета длины волны де Бройля частиц
  title: De Broglie Project API
  version: "1.0"
paths:
...
  /requestdebrogliecalculations:
    get:
      description: Возвращает список заявок на расчет с возможностью фильтрации
      parameters:
      - description: Статус заявки
        in: query
        name: status
        type: string
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/ds.RequestDeBroglieCalculation'
            type: array
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка заявок
      tags:
      - RequestDeBroglieCalculations
  /requestdebrogliecalculations/{id}:
    delete:
      description: Удаляет заявку на расчет
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление заявки
      tags:
      - RequestDeBroglieCalculations
    get:
      description: Возвращает детальную информацию о заявке с расчетами
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Детальная информация о заявке
          schema:
            $ref: '#/definitions/handler.deBroglieRequestDetailResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получение заявки по ID
      tags:
      - RequestDeBroglieCalculations
    put:
      consumes:
      - application/json
      description: Обновляет информацию о заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.RequestDeBroglieCalculation'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление заявки
      tags:
      - RequestDeBroglieCalculations
  /requestdebrogliecalculations/{id}/complete:
    put:
      consumes:
      - application/json
      description: Одобряет или отклоняет заявку (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Решение модератора
        in: body
        name: request
        required: true
        schema:
          properties:
            approve:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешное завершение
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Завершение заявки модератором
      tags:
      - RequestDeBroglieCalculations
  /requestdebrogliecalculations/{id}/form:
    put:
      description: Переводит черновик заявки в статус "сформирована"
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное формирование
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный ID или статус заявки
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Формирование заявки
      tags:
      - RequestDeBroglieCalculations
  /requestdebrogliecalculations/debrogliecart:
    get:
      description: Возвращает информацию о черновике заявки пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Информация о черновике
          schema:
            $ref: '#/definitions/handler.deBroglieDraftInfoResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Информация о черновике заявки
      tags:
      - RequestDeBroglieCalculations
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"